requirements.txt
cryptography
apache-airflow>=2.8.0

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0

# ETL and data fetching
requests>=2.31.0
python-dotenv>=1.0.0
polygon-api-client>=1.14.0

# Technical indicators
pandas-ta>=0.4.0

# Optional: LLM sentiment analysis (install with: pip install transformers)
# transformers>=4.30.0

# Testing and code quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
pre-commit>=3.5.0

